+++ changelog.md
@@ -17,6 +17,7 @@
  - Support KDE Breeze Dark theme (#407)
  - Use new Kodi XML syntax for movie set names (#554)
  - Add context menu in movie duplicate view (#591)
+ - Add AEBN genre option (#590)
  - Scrape TV show tags when using IMDb
  - Detect duplicate movies
  - Create subdirectories